move runToolWithStreaming in @lib/api
