+++ front/temporal/agent_loop/activities/run_tool.ts
@@ -1,6 +1,6 @@
 import assert from "assert";
 
-import { runToolWithStreaming } from "@app/lib/actions/mcp";
+import { MCPActionType } from "@app/lib/actions/mcp";
 import type { AuthenticatorType } from "@app/lib/auth";
 import { Authenticator } from "@app/lib/auth";
 import { AgentMCPActionResource } from "@app/lib/resources/agent_mcp_action_resource";
@@ -8,6 +8,7 @@ import { updateResourceAndPublishEvent } from "@app/temporal/agent_loop/activiti
 import { buildActionBaseParams } from "@app/temporal/agent_loop/lib/action_utils";
 import type { ToolExecutionResult } from "@app/temporal/agent_loop/lib/deferred_events";
 import { sliceConversationForAgentMessage } from "@app/temporal/agent_loop/lib/loop_utils";
+import { runToolWithStreaming } from "@app/temporal/agent_loop/lib/run_tool_streaming";
 import type { ModelId } from "@app/types";
 import { assertNever } from "@app/types";
 import type { RunAgentArgs } from "@app/types/assistant/agent_run";
@@ -53,10 +54,7 @@ export async function runToolActivity(
       step: step + 1,
     });
 
-  const action = await AgentMCPActionResource.fetchByModelIdWithAuth(
-    auth,
-    actionId
-  );
+  const action = await AgentMCPActionResource.fetchById(auth, actionId);
   assert(action, "Action not found");
 
   const mcpServerId = action.toolConfiguration.toolServerId;
@@ -77,6 +75,12 @@ export async function runToolActivity(
     agentConfiguration,
     agentMessage,
     conversation,
+    mcpAction: new MCPActionType({
+      ...actionBaseParams,
+      id: action.id,
+      output: null,
+      type: "tool_action",
+    }),
   });
 
   for await (const event of eventStream) {
